added fusioncatcher db download to setup.sh

+++ setup.sh
@@ -5,7 +5,7 @@ SCRIPT_PATH=$(
   pwd -P
 )
 
-readonly VALID_TASKS=("all db_install db_setup tools_install tools_setup ref example")
+readonly VALID_TASKS=("all db_install db_setup tools_install tools_setup fusioncatcherdb ref example")
 
 function join_by { local IFS="$1"; shift; echo "$*"; }
 
@@ -188,14 +188,25 @@ function setup_tool_fusioncatcher() {
 
   mkdir -p "${DIR_TOOLS}/fusioncatcher/data"
   cd "${DIR_TOOLS}/fusioncatcher/data" || exit 1
-  wget http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.aa
-  wget http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.ab
-  wget http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.ac
-  wget http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.ad
-  cat human_v102.tar.gz.* | tar xz
+  wget --no-check-certificate http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.aa -O human_v102.tar.gz.aa
+  wget --no-check-certificate http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.ab -O human_v102.tar.gz.ab
+  wget --no-check-certificate http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.ac -O human_v102.tar.gz.ac
+  wget --no-check-certificate http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.tar.gz.ad -O human_v102.tar.gz.ad
+  wget --no-check-certificate http://sourceforge.net/projects/fusioncatcher/files/data/human_v102.md5 -O human_v102.md5
+  md5sum -c human_v102.md5
+  if [ "$?" -ne "0" ]; then
+    echo -e "\n\n\n\033[33;7m   ERROR: The downloaded files from above have errors! MD5 checksums do not match! Please, download them again or re-run this script again!   \033[0m\n"
+    exit 1
+  fi
+
+  cat human_v102.tar.gz.* > human_v102.tar.gz
+  rm -f human_v102.tar.gz.*
+
+  tar -xzf human_v102.tar.gz
   ln -s human_v102 current
 
   rm -f human_v102.tar.gz
+  rm -f human_v102.md5
   echo "done"
 }
